volume_1/chapter_2/anthropic/requirements.txt
anthropic>=0.31.0
python-dotenv>=1.0.0
